SOLR-15904: Move SQL support into new sql module
