+++ gradle/validation/spotless.gradle
@@ -53,6 +53,7 @@ configure(project(":solr").subprojects) { prj ->
           case ":solr:modules:hdfs":
           case ":solr:modules:langid":
           case ":solr:modules:scripting":
+          case ":solr:modules:sql":
           case ":solr:core":
           case ":solr:solrj":
           case ":solr:solr-ref-guide":
