+++ settings.gradle
@@ -42,6 +42,7 @@ include "solr:modules:langid"
 include "solr:modules:ltr"
 include "solr:modules:s3-repository"
 include "solr:modules:scripting"
+include "solr:modules:sql"
 include "solr:webapp"
 include "solr:benchmark"
 include "solr:test-framework"
