+++ solr/CHANGES.txt
@@ -109,6 +109,8 @@ New Features
 
 * SOLR-13989: Move Hadoop Authentication support to a new hadoop-auth module (Kevin Risden)
 
+* SOLR-15904: Move SQL support to new sql module (Kevin Risden)
+
 Improvements
 ----------------------
 * LUCENE-8984: MoreLikeThis MLT is biased for uncommon fields (Andy Hind via Anshum Gupta)
