+++ solr/modules/sql/src/java/org/apache/solr/handler/sql/LimitStream.java
