+++ solr/modules/sql/src/java/org/apache/solr/handler/sql/package-info.java
@@ -16,6 +16,6 @@
  */
 
 /**
- * Classes related to Apache Calcite implementation in {@link org.apache.solr.handler.SQLHandler}
+ * Classes related to Apache Calcite implementation in {@link org.apache.solr.handler.sql.SQLHandler}
  */
 package org.apache.solr.handler.sql;
