+++ solr/modules/sql/src/test-files/log4j2.xml
@@ -32,6 +32,7 @@
     <Logger name="org.apache.zookeeper" level="WARN"/>
     <Logger name="org.apache.directory" level="WARN"/>
     <Logger name="org.eclipse.jetty" level="INFO"/>
+    <Logger name="org.apache.calcite" level="INFO"/>
 
     <Root level="INFO">
       <AppenderRef ref="STDERR"/>
