+++ solr/modules/sql/src/test-files/solr/configsets/sql/conf/solrconfig.xml
