+++ solr/test-framework/src/java/org/apache/solr/schema/SortableBinaryField.java
