+++ versions.lock
@@ -35,7 +35,7 @@ com.google.cloud:google-cloud-storage:2.2.3 (2 constraints: 6e1c99f3)
 com.google.code.gson:gson:2.8.9 (6 constraints: f25833c8)
 com.google.errorprone:error_prone_annotations:2.10.0 (4 constraints: 3038d4ba)
 com.google.guava:failureaccess:1.0.1 (2 constraints: f9199e37)
-com.google.guava:guava:31.0.1-jre (19 constraints: 722c2e36)
+com.google.guava:guava:31.0.1-jre (17 constraints: 850e5ccb)
 com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava (2 constraints: 4b35b0a0)
 com.google.http-client:google-http-client:1.41.0 (11 constraints: 17bf6b7e)
 com.google.http-client:google-http-client-apache-v2:1.41.0 (2 constraints: af201c75)
